# back-end/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "BED EP1 API - Martinus Nordgård",
    "description": "Documentation for the BED Exam Project API. Contains all nessecary information about how to use this API successfully. Documentation <del>automatically</del> partially generated by the swagger-autogen module and partially written by the Author."
  },
  "host": "localhost:3001",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Retrieve all users",
        "description": "Fetch all users from the database.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved users.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Users found"
                    },
                    "users": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 2
                          },
                          "Username": {
                            "type": "string",
                            "example": "user123"
                          },
                          "EncryptedPassword": {
                            "type": "string",
                            "example": "8e85b2d8dbd679716888c37d262bdb1f7cd84e7e3c28339d971fac741650d98c"
                          },
                          "Salt": {
                            "type": "string",
                            "example": "215d32304553f246781f52d387790a4a"
                          },
                          "Email": {
                            "type": "string",
                            "example": "Joe@gmail.com"
                          },
                          "FirstName": {
                            "type": "string",
                            "example": "Joe"
                          },
                          "LastName": {
                            "type": "string",
                            "example": "Smith"
                          },
                          "Address": {
                            "type": "string",
                            "example": "123 Street, City"
                          },
                          "Telephone": {
                            "type": "string",
                            "example": "242424"
                          },
                          "MembershipId": {
                            "type": "number",
                            "example": 1
                          },
                          "RoleId": {
                            "type": "number",
                            "example": 2
                          },
                          "createdAt": {
                            "type": "string",
                            "example": "2025-03-06T14:58:26.000Z"
                          },
                          "updatedAt": {
                            "type": "string",
                            "example": "2025-03-06T14:58:26.000Z"
                          },
                          "RoleName": {
                            "type": "string",
                            "example": "User"
                          },
                          "MembershipName": {
                            "type": "string",
                            "example": "Bronze"
                          },
                          "TotalItemsPurchased": {
                            "type": "string",
                            "example": "0"
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "statuscode": {
                  "type": "number",
                  "example": 404
                },
                "execution": {
                  "type": "string",
                  "example": "database called successfully"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "No users found"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error while fetching users.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "statuscode": {
                  "type": "number",
                  "example": 500
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "An error occurred while fetching users"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Retrieve all users",
        "description": "Fetch all users from the database.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved users.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "User found"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 2
                        },
                        "Username": {
                          "type": "string",
                          "example": "user123"
                        },
                        "EncryptedPassword": {
                          "type": "string",
                          "example": "8e85b2d8dbd679716888c37d262bdb1f7cd84e7e3c28339d971fac741650d98c"
                        },
                        "Salt": {
                          "type": "string",
                          "example": "215d32304553f246781f52d387790a4a"
                        },
                        "Email": {
                          "type": "string",
                          "example": "Joe@gmail.com"
                        },
                        "FirstName": {
                          "type": "string",
                          "example": "Joe"
                        },
                        "LastName": {
                          "type": "string",
                          "example": "Smith"
                        },
                        "Address": {
                          "type": "string",
                          "example": "123 Street, City"
                        },
                        "Telephone": {
                          "type": "string",
                          "example": "242424"
                        },
                        "MembershipId": {
                          "type": "number",
                          "example": 1
                        },
                        "RoleId": {
                          "type": "number",
                          "example": 2
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2025-03-06T14:58:26.000Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "example": "2025-03-06T14:58:26.000Z"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "statuscode": {
                  "type": "number",
                  "example": 404
                },
                "execution": {
                  "type": "string",
                  "example": "database called successfully"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error while fetching users.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "statuscode": {
                  "type": "number",
                  "example": 500
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "An error occurred while fetching user"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/users/{id}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete a specific user",
        "description": "Delete a user by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "The ID of the user to delete."
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the user.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User deleted successfully"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error while deleting user.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "statuscode": {
                  "type": "number",
                  "example": 500
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "An error occurred while deleting the user"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update a user",
        "description": "Update the details of an existing user.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "The ID of the user to update."
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "User details to update.",
            "schema": {
              "type": "object",
              "properties": {
                "firstname": {
                  "type": "string",
                  "example": "John"
                },
                "lastname": {
                  "type": "string",
                  "example": "Doe"
                },
                "username": {
                  "type": "string",
                  "example": "johndoe"
                },
                "email": {
                  "type": "string",
                  "example": "johndoe@example.com"
                },
                "address": {
                  "type": "string",
                  "example": "123 Main St"
                },
                "telephone": {
                  "type": "string",
                  "example": "123456789"
                },
                "membershipId": {
                  "type": "number",
                  "example": 1
                },
                "roleId": {
                  "type": "number",
                  "example": 2
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated user details.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User updated successfully"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error while updating user details.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "statuscode": {
                  "type": "number",
                  "example": 500
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "An error occurred while updating the user"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  }
}